+++ src/app.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
 import { ApiV1Module } from './api-v1/api-v1.module';
+import { MetricsModule } from './metrics/metrics.module';
 import { MongoModule } from './mongo/mongo.module';
 
 @Module({
-  imports: [MongoModule, ApiV1Module],
+  imports: [MetricsModule, MongoModule, ApiV1Module],
 })
 export class AppModule {}